권한 처리에 대한 작업중
- 파일 열기는 아직 작업중

+++ server/routes/asset/file-body/delete-remove.js
@@ -8,7 +8,8 @@
 import { success, error } from '../../output.js'
 import { connect, disconnect, tables, getItem, removeItem } from '../../../libs/db.js'
 import { checkAuthorization } from '../../../libs/token.js'
-import { removeFile } from '../../../libs/service.js'
+import { removeFile, checkAssetOwner } from '../../../libs/service.js'
+import { ownerModes } from '../../../../global/consts.js'
 import ServiceError from '../../../libs/ServiceError.js'
 
 export default async (req, res) => {
@@ -22,7 +23,10 @@ export default async (req, res) => {
     // connect db
     connect({ readwrite: true })
     // check auth
-    checkAuthorization(req.headers.authorization)
+    const auth = checkAuthorization(req.headers.authorization)
+
+    // check owner
+    checkAssetOwner(ownerModes.ASSET, auth.id, assetId)
 
     // get file data
     const file = getItem({
@@ -30,6 +34,7 @@ export default async (req, res) => {
       where: `id = $id`,
       values: { '$id': fileId },
     })
+    if (!file?.data) throw new ServiceError('파일 데이터가 없습니다.')
 
     // remove data
     if (file?.data?.path) removeFile(file.data.path)
